Update README.md
Remove outdated links. Steve has (on his podcast and one the forum, maybe not really on the main site...) explained that the links to subpages under grc.com/sqrl were removed because he was working on a "PDF version" of the important information. And now the main page has 4 documents. Likely still some parts missing there, that's when the forum comes in handy: https://sqrl.grc.com/